May required updates and check changes.
This changes updates the spellcheck suppression entries in the files that require them. It fixes some of the problems pointed out by the Shellcheck linter. It also splits-up the aliases into atomic definitions to ease its management. It also updated the GitHub profile.d definition and prefixed some internal functions with ‘__’ to sign them as internal.
There were some addition of sanitize functions to deal with path lists, to make sure that there are no repeating entries in the list, nor that there are invalid entries too.