Make sure the PATH is properly set
This change adds an internal function to perform the first PATH setting
if your system doesn't have a global command to do so. Otherwise it uses
the system command and sets it accordingly.
It also fixes some bugs found in the sanitizer internal function, as
well as other bugs in the internal path manipulation functions.